resume_project/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
openai==1.3.8
pydantic==2.5.1
aiofiles==23.2.1
python-dotenv==1.0.0